requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
autograd==1.8.0
autoray==0.7.1
babel==2.16.0
beautifulsoup4==4.12.3
bitarray==3.0.0
bleach==6.1.0
cached-property==2.0.1
cachetools==6.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chex==0.1.89
ckzg==2.0.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.3.8
eth-account==0.13.5
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-tester==0.12.1b1
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
etils==1.12.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.14.0
h5py==3.11.0
hexbytes==1.3.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
lru-dict==1.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
namex==0.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.2
opt_einsum==3.4.0
optax==0.2.4
optree==0.16.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pbr==6.1.0
PennyLane==0.41.0
PennyLane_Lightning==0.41.1
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
py-ecc==7.0.1
py-evm==0.10.1b2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==4.0.0.dev0
pyunormalize==16.0.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qiskit==1.2.0
qtealeaves==1.2.30
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rlp==4.1.0
rpds-py==0.20.0
rustworkx==0.15.1
scikit-learn==1.7.0rc1
scipy==1.16.0rc2
scipy-openblas32==0.3.29.265.0
semantic-version==2.10.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
stevedore==5.3.0
symengine==0.11.0
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tomlkit==0.13.2
toolz==1.0.0
torch==2.7.1
tornado==6.4.1
traitlets==5.14.3
trie==3.1.0
types-python-dateutil==2.9.0.20240821
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
web3==7.8.0
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
